pyproject.toml
[project]
name = "google-adk-test"
version = "0.1.0"
description = "智能投资"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.5.0",
    "litellm>=1.74.0",
    "tavily-python>=0.7.8",
]
